+++ docs/release-notes.md
@@ -7,10 +7,12 @@
   _If needed, you can update to SMAPI 3.16.0 first and then install the latest version._
 -->
 
-## Upcoming release
+## 3.18.6
+Released 05 October 2023 for Stardew Valley 1.5.6 or later.
+
 * For players:
+  * Fixed SpriteMaster compatibility in SMAPI 3.18.5+ with temporary workaround.
   * Fixed `player_add` and `list_items` console commands not including Pickled Ginger, and returning Honey instead of Wild Honey _(in Console Commands)_.
-  * Fixed compatibility with SpriteMaster in SMAPI 3.18.5.
 
 * For mod authors:
   * Added asset propagation for `LooseSprites/chatBox` and `LooseSprites/emojis`.